Refactoring some code!
Hi all,

I’ve just went ahead and started refactoring some of the code so it is
more legible for us. I have also been experimenting with making the
Maps available online but this is looking to be a bit more tricky than
I thought so I have excluded those changes.